+++ platform-tools-sdk/cargo-build-sbf/src/main.rs
@@ -73,7 +73,7 @@ impl Default for Config<'_> {
             verbose: false,
             workspace: false,
             jobs: None,
-            arch: "sbfv1",
+            arch: "v0",
         }
     }
 }
@@ -624,7 +624,13 @@ fn build_solana_package(
         .cloned()
         .unwrap_or_else(|| target_directory.join("deploy"));
 
-    let target_build_directory = target_directory.join("sbf-solana-solana").join("release");
+    let target_triple = if config.arch == "v0" {
+        "sbpf-solana-solana".to_string()
+    } else {
+        format!("sbpf{}-solana-solana", config.arch)
+    };
+
+    let target_build_directory = target_directory.join(&target_triple).join("release");
 
     env::set_current_dir(root_package_dir).unwrap_or_else(|err| {
         error!(
@@ -705,10 +711,9 @@ fn build_solana_package(
             exit(1);
         });
     }
-    let target = "sbf-solana-solana";
 
     if config.no_rustup_override {
-        check_solana_target_installed(target);
+        check_solana_target_installed(&target_triple);
     } else {
         link_solana_toolchain(config);
         // RUSTC variable overrides cargo +<toolchain> mechanism of
@@ -734,7 +739,10 @@ fn build_solana_package(
     env::set_var("OBJDUMP", llvm_bin.join("llvm-objdump"));
     env::set_var("OBJCOPY", llvm_bin.join("llvm-objcopy"));
 
-    let cargo_target = "CARGO_TARGET_SBF_SOLANA_SOLANA_RUSTFLAGS";
+    let cargo_target = format!(
+        "CARGO_TARGET_{}_RUSTFLAGS",
+        target_triple.to_uppercase().replace("-", "_")
+    );
     let rustflags = env::var("RUSTFLAGS").ok().unwrap_or_default();
     if env::var("RUSTFLAGS").is_ok() {
         warn!(
@@ -743,7 +751,7 @@ fn build_solana_package(
         );
         env::remove_var("RUSTFLAGS")
     }
-    let target_rustflags = env::var(cargo_target).ok();
+    let target_rustflags = env::var(&cargo_target).ok();
     let mut target_rustflags = Cow::Borrowed(target_rustflags.as_deref().unwrap_or_default());
     target_rustflags = Cow::Owned(format!("{} {}", &rustflags, &target_rustflags));
     if config.remap_cwd && !config.debug {
@@ -753,17 +761,14 @@ fn build_solana_package(
         // Replace with -Zsplit-debuginfo=packed when stabilized.
         target_rustflags = Cow::Owned(format!("{} -g", &target_rustflags));
     }
-    if config.arch == "sbfv2" {
-        target_rustflags = Cow::Owned(format!("{} -C target_cpu=sbfv2", &target_rustflags));
-    }
     if let Cow::Owned(flags) = target_rustflags {
-        env::set_var(cargo_target, flags);
+        env::set_var(&cargo_target, flags);
     }
     if config.verbose {
         debug!(
             "{}=\"{}\"",
             cargo_target,
-            env::var(cargo_target).ok().unwrap_or_default(),
+            env::var(&cargo_target).ok().unwrap_or_default(),
         );
     }
 
@@ -773,10 +778,7 @@ fn build_solana_package(
         cargo_build_args.push("+solana");
     };
 
-    cargo_build_args.append(&mut vec!["build", "--release", "--target", target]);
-    if config.arch == "sbfv2" {
-        cargo_build_args.push("-Zbuild-std=std,panic_abort");
-    }
+    cargo_build_args.append(&mut vec!["build", "--release", "--target", &target_triple]);
     if config.no_default_features {
         cargo_build_args.push("--no-default-features");
     }
@@ -911,7 +913,10 @@ fn build_solana_package(
             }
         }
 
-        check_undefined_symbols(config, &program_so);
+        if config.arch != "v3" {
+            // SBPFv3 shall not have any undefined syscall.
+            check_undefined_symbols(config, &program_so);
+        }
 
         info!("To deploy this program:");
         info!("  $ solana program deploy {}", program_so.display());
@@ -1138,8 +1143,8 @@ fn main() {
         .arg(
             Arg::new("arch")
                 .long("arch")
-                .possible_values(["sbfv1", "sbfv2"])
-                .default_value("sbfv1")
+                .possible_values(["v0", "v1", "v2", "v3"])
+                .default_value("v0")
                 .help("Build for the given target architecture"),
         )
         .get_matches_from(args);