+++ CHANGELOG.md
@@ -14,7 +14,9 @@ Release channels have their own copy of this changelog:
 
 <a name="edge-channel"></a>
 ## 2.3.0 - Unreleased
-
+* Changes:
+  * Platform tools SDK:
+    * `cargo-build-sbf` and `cargo-test-sbf` now accept `v0`, `v1`, `v2` and `v3` for the `--arch` argument. These parameters specify the SBPF version to build for.
 
 ## 2.2.0
 * Breaking: